+++ packages/ng-primitives/date-time/src/native-date-adapter/native-date-adapter.spec.ts
@@ -1,0 +1,4 @@
+import { dateAdapterTests } from 'ng-primitives/date-time/testing';
+import { NgpNativeDateAdapter } from './native-date-adapter';
+
+dateAdapterTests(NgpNativeDateAdapter);