+++ packages/ng-primitives/listbox/src/index.ts
@@ -1,3 +1,4 @@
+export { NgpListboxHeader } from './listbox-header/listbox-header';
 export { NgpListboxOption } from './listbox-option/listbox-option';
 export { NgpListboxSection } from './listbox-section/listbox-section';
 export { NgpListboxTrigger } from './listbox-trigger/listbox-trigger';